requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# LLM integration (local by default)
ollama==0.1.6

# Optional: OpenAI-compatible endpoints
openai==1.10.0

# Testing
pytest==7.4.3
httpx>=0.25.2,<0.26.0

# Optional for enhanced metrics
textblob==0.17.1
